fix: withTail 方法参数类型由 long 改为 Integer

+++ src/main/java/com/nexa/flowops/util/DockerUtil.java
@@ -26,7 +26,7 @@ public class DockerUtil {
         }
     }
 
-    public String getContainerLog(String containerName, long tail) {
+    public String getContainerLog(String containerName, Integer tail) {
         try {
             return dockerClient.logContainerCmd(containerName)
                     .withTail(tail)
